requirements.txt

pandas==2.3.3
scikit-learn==1.7.2
tqdm==4.67.3
timm==1.0.26
streamlit==1.33.0
numpy==1.26.4
pillow==10.3.0
matplotlib==3.8.4
torch==2.2.2
torchvision==0.17.2